pyproject.toml
[project]
name = "meetingwhisperer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "google-api-python-client>=2.166.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "mcp[cli]>=1.6.0",
    "openai>=1.72.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
